pyproject.toml
[project]
name = "qa-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainit>=0.4.0",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
]
